Agent_data/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
babel==2.18.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
courlan==1.3.2
dateparser==1.2.2
exceptiongroup==1.3.1
fastapi==0.128.8
feedparser==6.0.12
h11==0.16.0
htmldate==1.9.4
idna==3.11
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
pydantic==2.12.5
pydantic_core==2.41.5
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2026.1.15
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
starlette==0.49.3
tld==0.13.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
